requirements.txt
aiohttp~=3.8.3
beautifulsoup4~=4.11.2
discord.py~=2.1.0
aiohttp-client-cache~=0.8.1
SQLAlchemy~=2.0.2